# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Anikultura API Docs",
        "description": "Documentation for the Anikultura REST API",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/",
            "description": "This anikultura server"
        },
        {
            "url": "https://ani-be-stg.herokuapp.com/api/",
            "description": "Anikultura staging server"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "description": "Login to acquire an authentication token",
                "operationId": "36f0bcf86738f6ec62f6ca166fa43165",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "user1"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login with returned authentication token",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors occured",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid login credentials",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Already logged in",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "auth"
                ],
                "description": "Logout from current authenticated session",
                "operationId": "27cd4e2d48a3f73d894c4834e4b82604",
                "responses": {
                    "200": {
                        "description": "Successful logout",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/crops/seed-allocation": {
            "get": {
                "tags": [
                    "crops"
                ],
                "description": "Get the allocated seeds of the logged in farmer",
                "operationId": "f040454776564edc6e13b540b930e3e9",
                "responses": {
                    "200": {
                        "description": "The allocated seeds",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/crops": {
            "get": {
                "tags": [
                    "crops"
                ],
                "description": "Get the current crops of the logged in farmer",
                "operationId": "06d09797e7a39bacc8fd0ad9ea6df07e",
                "responses": {
                    "200": {
                        "description": "The current crops",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/crops/current-seed-stage": {
            "post": {
                "tags": [
                    "crops"
                ],
                "description": "Get the current seed stage of the logged in farmer",
                "operationId": "442ec1357ece8d70c38f05f9512919b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "farmlandId": {
                                        "type": "integer",
                                        "format": "int",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The current seed stage",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/crops/next-seed-stage": {
            "post": {
                "tags": [
                    "crops"
                ],
                "description": "Get the next seed stage of the logged in farmer",
                "operationId": "7ef59a557df3ba6691646184b57242d6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "farmlandId": {
                                        "type": "integer",
                                        "format": "int",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The next seed stage",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/farmers/language": {
            "get": {
                "tags": [
                    "farmers"
                ],
                "description": "Get the current language preference of the logged in farmer",
                "operationId": "b278227a79c4aeaa9fd4266fd6d2d47e",
                "responses": {
                    "200": {
                        "description": "The language preference",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "farmers"
                ],
                "description": "Update the language preference of the logged in farmer",
                "operationId": "8ec40af7b9c94e5864822bcdebece226",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "language": {
                                        "type": "string",
                                        "example": "en"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful update",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/farmers/tutorial": {
            "get": {
                "tags": [
                    "farmers"
                ],
                "description": "Get the current tutorial state of the logged in farmer",
                "operationId": "ab0651cffffbbafdc4395ca8e7d92df8",
                "responses": {
                    "200": {
                        "description": "The tutorial state",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "farmers"
                ],
                "description": "Update the tutorial state of the logged in farmer",
                "operationId": "a4b592b600ba3b3f43b17bf773d39845",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tutorialDone": {
                                        "type": "boolean",
                                        "example": "false"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful update",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/farmer-reports/{farmlandId}": {
            "get": {
                "tags": [
                    "farmer-reports"
                ],
                "description": "Get the submitted farmer reports of the logged in farmer",
                "operationId": "8565cb07b551ec43c674ae8b21ff2810",
                "parameters": [
                    {
                        "name": "farmlandId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The submitted farmer reports",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/farmer-reports": {
            "post": {
                "tags": [
                    "farmer-reports"
                ],
                "description": "Submit a farming report with logged in farmer",
                "operationId": "457a393b51e876064179a323dc93e50c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "farmerReport": {
                                        "properties": {
                                            "farmlandId": {
                                                "type": "integer",
                                                "format": "int",
                                                "example": "1"
                                            },
                                            "cropId": {
                                                "type": "integer",
                                                "format": "int",
                                                "example": "1"
                                            },
                                            "volumeKg": {
                                                "type": "number",
                                                "format": "double",
                                                "example": "10.23"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The created farmer report",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors occured",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/farmer-reports/{farmerReportId}/upload": {
            "post": {
                "tags": [
                    "farmer-reports"
                ],
                "description": "Upload an image to farmer report",
                "operationId": "7ecdc8be22a79d90a7dd225a0317143d",
                "parameters": [
                    {
                        "name": "farmerReportId",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "description": "proof image to upload",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Upload success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors occured",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/farmlands": {
            "get": {
                "tags": [
                    "farmlands"
                ],
                "description": "Get the associated farmlands of the logged in farmer",
                "operationId": "0fa449de791367f98f88a3fda6396eaf",
                "responses": {
                    "200": {
                        "description": "The associated farmlands",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "auth",
            "description": "Endpoints related with authentication such as login and logout."
        },
        {
            "name": "crops",
            "description": "Endpoints related with crops such as retrieving current seed stage and allocated seeds."
        },
        {
            "name": "farmlands",
            "description": "Endpoints related with farmlands such as what farmland a farmer belongs in."
        },
        {
            "name": "farmers",
            "description": "Endpoints related with farmers such as updating tutorial state."
        },
        {
            "name": "farmer-reports",
            "description": "Endpoints related with farmer reports such as submitting and retrieving their history."
        }
    ]
}